requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pyodbc
passlib[bcrypt]
python-jose
pytest